drivers: Work on usb/stm32_hc, Extract f4 specific part
