+++ src/drivers/usb/stm32_hc/stm32f4cube_hc.c
@@ -8,11 +8,13 @@
 
 #include <util/log.h>
 
-#include "stm32_hc.h"
+#include <drivers/usb/hc/stm32_hc_conf.h>
 
 #include <stm32f4xx.h>
 #include <stm32f4xx_hal.h>
 #include <stm32f4xx_hal_hcd.h>
+
+extern HCD_HandleTypeDef stm32_hcd_handler;
 
 void HAL_HCD_MspInit(HCD_HandleTypeDef *hhcd) {
 	GPIO_InitTypeDef GPIO_InitStructure;
@@ -36,19 +38,18 @@ void HAL_HCD_MspInit(HCD_HandleTypeDef *hhcd) {
 	HAL_GPIO_Init(GPIOA, &GPIO_InitStructure);
 
 	GPIO_InitStructure.Pin = GPIO_PIN_10;
-    GPIO_InitStructure.Mode = GPIO_MODE_AF_OD;
-    GPIO_InitStructure.Pull = GPIO_PULLUP;
-    GPIO_InitStructure.Alternate = GPIO_AF10_OTG_FS;
-    HAL_GPIO_Init(GPIOA, &GPIO_InitStructure);
+	GPIO_InitStructure.Mode = GPIO_MODE_AF_OD;
+	GPIO_InitStructure.Pull = GPIO_PULLUP;
+	GPIO_InitStructure.Alternate = GPIO_AF10_OTG_FS;
+	HAL_GPIO_Init(GPIOA, &GPIO_InitStructure);
 
-    GPIO_InitStructure.Pin = STM32_HC_VBUS_PIN;
-    GPIO_InitStructure.Mode = GPIO_MODE_OUTPUT_PP;
-    GPIO_InitStructure.Pull = GPIO_NOPULL;
-    HAL_GPIO_Init(STM32_HC_VBUS_PORT, &GPIO_InitStructure);
+	GPIO_InitStructure.Pin = STM32_HC_VBUS_PIN;
+	GPIO_InitStructure.Mode = GPIO_MODE_OUTPUT_PP;
+	GPIO_InitStructure.Pull = GPIO_NOPULL;
+	HAL_GPIO_Init(STM32_HC_VBUS_PORT, &GPIO_InitStructure);
 #else
 #error "don't support platform"
 #endif
 	/* Enable USB clock  */
 	__HAL_RCC_USB_OTG_FS_CLK_ENABLE();
 }
-
